fix: 修复配置文件没有有效迁移的问题
