+++ util/general_utils.py
@@ -15,7 +15,6 @@ from PIL import Image, ImageDraw, ImageFont
 from type.types import GlobalObject
 from SparkleLogging.utils.core import LogManager
 from logging import Logger
-from collections import defaultdict
 
 logger: Logger = LogManager.GetLogger(log_name='astrbot-core')
 
@@ -421,21 +420,6 @@ def create_markdown_image(text: str):
         return p
     except Exception as e:
         raise e
-
-
-def try_migrate_config():
-    '''
-    将 cmd_config.json 迁移至 data/cmd_config.json
-    '''
-    if os.path.exists("cmd_config.json"):
-        with open("cmd_config.json", "r", encoding="utf-8-sig") as f:
-            data = json.load(f)
-        with open("data/cmd_config.json", "w", encoding="utf-8-sig") as f:
-            json.dump(data, f, indent=2, ensure_ascii=False)
-        try:
-            os.remove("cmd_config.json")
-        except Exception as e:
-            pass
 
 
 def get_local_ip_addresses():
